+++ docs/changelog.md
@@ -2,10 +2,10 @@
 
 ## 1.2.0
 #### qqrobot-spring-boot-starter 1.2.0
-- TesterProperties的webPrefix转为静态常量。
+- `TesterProperties`的webPrefix转为静态常量。
 - 修复Tester Web界面的at功能bug。
-- 为Robot封装通用的ConsoleWindow。
-- 解决AdminLoginInterceptor潜在的错误拦截问题。
+- 为Robot封装通用的`ConsoleWindow`。
+- 解决`AdminLoginInterceptor`潜在的错误拦截问题。
 - 优化使用记录界面的表格列宽度。
 - 使用函数式接口更新会话创建方式。